Added `tree` option to default (linux) and macOS kill methods
When `tree` option is enabled the process specified by the input, and its descendants, will be killed.
To kill the whole tree it has been added and used the module `tree-kill`.
To handle a process name as input, which can match to multiple processes, it has
been added the function `getPidsFromInput`, which uses `ps-list` to get all the
pids by the proces name.