Tentative fixes for windows test for kill tree
Various errors for pathing for fixtures has been corrected, mostly caused
by `child_process.spawn`.
On Windows it's not possible to rename a process in the task list via JS,
so the test to kill tree by name has been disabled for Windows.
On Windows, to taskkill a tree `force` must be enabled.