Kill a process tree using `pidtree`
AFAIK macOS and Linux can use the same technique for killing a process tree
and a single process via `process.kill`, as such their implementation have
been merged into `defaultKill`;

I've removed missingBinaryError, because it was used for checking if the
commands `kill` or `killall` were available on the host, but now it just
use `process.kill`.

`options.tree` has been defaulted from the start because it's now used for
both the kill methods implementations (Linux/macOS and Windows).

The type of `input` in `handleKill` is now more important and the method
have been streamlined; if the input is a:

- process name (string), it gathers all the pids, similar to what
`killAll` would do, and filter all the processes by it's name, excluding
the `fkill` process itself;
- process PID (number), it checks the existance of the process and proceed
to kill it.

As per the above changes the check for the existance of a process have
been moved from the catch and duplicated accordingly to the type of the
input.